refactor(pci-rancher): rename and move pci rancher app file base ref: TAPC-987
