--- packages/manager/apps/pci-rancher/src/global.css
@@ -1,1 +1,0 @@
-@tailwind utilities;
