+++ packages/manager/apps/pci-rancher/src/index.html
@@ -18,6 +18,6 @@
             </strong>
         </noscript>
         <div id="root"></div>
-        <script type="module" src="./index.tsx"></script>
+        <script type="module" src="./main.tsx"></script>
     </body>
 </html>
