+++ packages/manager/apps/pci-rancher/src/index.scss
@@ -1,3 +1,5 @@
+@tailwind utilities;
+
 .border-ods-primary-100 {
   border-color: var(--ods-color-primary-100);
 }
