+++ packages/manager/apps/pci-rancher/src/main.tsx
@@ -7,7 +7,7 @@ import {
 import App from './App';
 import initI18n from './i18n';
 
-import './global.css';
+import './index.scss';
 
 import '@/vite-hmr';
