+++ packages/manager/apps/pci-rancher/src/routes/Router.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { RouterProvider, createHashRouter } from 'react-router-dom';
-import appRoutes from '@/routes';
+import appRoutes from '@/routes/routes';
 
 const router = createHashRouter(appRoutes);
